backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
groq>=0.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
